adding a readme

+++ math/probability/poisson.py
@@ -12,7 +12,7 @@ class Poisson:
     Lambtha
     """
 
-    def __init__(self, data=None, lambtha=1):
+    def __init__(self, data=None, lambtha=1.0):
         if data is None:
             if lambtha <= 0:
                 raise ValueError("lambtha must be a positive value")
@@ -23,7 +23,7 @@ class Poisson:
             if len(data) < 2:
                 raise ValueError("data must contain multiple values")
 
-            self.lambtha = float(sum(data) / len(data))
+            self.lambtha = sum(data) / len(data)
             self.data = data
 
     def exp(self, x):
